requirements.txt
astroid==2.2.5
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.1
django-heroku==0.3.1
fake-useragent==0.1.11
gunicorn==19.9.0
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.16.3
pandas==0.24.2
pipenv==2018.11.26
psycopg2==2.8.2
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.1
regex==2019.5.25
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.25.3
virtualenv==16.6.0
virtualenv-clone==0.5.3
whitenoise==4.1.2
wrapt==1.11.1
